requirements.txt
requests>=2.31.0
numpy>=1.26.0
pandas>=2.1.0
pyarrow>=14.0.0
minio>=7.2.0
oracledb>=2.0.0
apache-airflow>=2.9.0
python-dotenv>=1.0.0
flake8>=7.0.0
pytest>=8.0.0
